fix defect 2 + split constants
