+++ js/constansts.js
@@ -1,0 +1,29 @@
+const COUNT_POSTS = 25;
+const NAMES_USER = [
+  'Иван',
+  'Себастьян',
+  'Мария',
+  'Кристоф',
+  'Виктор',
+  'Юлия',
+  'Люпита',
+  'Вашингтон',
+];
+
+const MESSAGE = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!',
+];
+
+const DESCRIPTION = [
+  'Описание: лес, море',
+  'Описание: кофе, чай',
+  'Описание: девушка, белка',
+  'Описание: волк, заяц',
+  'Описание: горы, что будоражат взор',
+];
+export {COUNT_POSTS, NAMES_USER, MESSAGE, DESCRIPTION};
